Add JWT header tests
