+++ lib/auth.js
@@ -1,7 +1,5 @@
 import GitHubProvider from 'next-auth/providers/github';
-import jwt from 'jsonwebtoken';
-
-const JWT_SECRET = process.env.NEXTAUTH_SECRET;
+import { getSubFromAuthHeader } from './token.js';
 
 export const authOptions = {
   providers: [
@@ -39,15 +37,4 @@ export const authOptions = {
   secret: process.env.NEXTAUTH_SECRET,
 };
 
-export function getSubFromAuthHeader(authHeader) {
-  if (!authHeader?.startsWith('Bearer ')) return null;
-
-  const token = authHeader.slice(7); // remove 'Bearer '
-  try {
-    const payload = jwt.verify(token, JWT_SECRET);
-    return payload?.sub || null;
-  } catch (err) {
-    console.warn('[JWT] Invalid token:', err);
-    return null;
-  }
-}
+export { getSubFromAuthHeader };
