+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "ws": "node server.js"
+    "ws": "node server.js",
+    "test": "node --test"
   },
   "dependencies": {
     "@apollo/server": "^4.12.1",
